Moving sql queries of link model to query object
