+++ app/serializers/link_serializer.rb
@@ -2,6 +2,6 @@ class LinkSerializer < ActiveModel::Serializer
   attributes :link_type, :title, :url, :keyword
 
   def keyword
-    object.search_result.keyword
+    object.keyword || object.search_result.keyword
   end
 end
